few small tweaks
